Add todo Re: def of lax nat transf r; Correct hocolim notation
Upon trying to understand how r = e^\op \wedge e is defined, I noticed
that I don't have a clue. -^\op doesn't seems to only be a 2-functor
CAT \to CAT^{\mathrm{co}}, where 2-arrows are turned around in the
codomain. Wrote an email to Morten about this, asking for halp.

Furthermore, I had made a mistake regarding the notation of hocolim
applied to a bimodule (F, G: F^e \to K): when noting the hocolim, the
e on top of F^e does not get noted!